+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
+The format is based on [Keep a Changelog](https://kSSeepachangelog.com/en/1.1.0/).
+
+## Unreleased
+
+### Security
+
+- Fix command injection vulnerability in `jdlib secrets` commands
 
 ## [0.0.7] - 2026-03-20